backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
supabase==2.3.4
sqlalchemy==2.0.27
python-dotenv==1.0.1
google-generativeai==0.4.1
pydantic[email]==2.6.1
websockets==12.0
bcrypt==4.1.2
httpx==0.27.0
